fixed the lexer issue where a single operator led to an error

+++ src/edu/kit/pp/minijava/Lexer.java
@@ -17,7 +17,7 @@ public class Lexer {
 
 	public Token next() throws IOException {
 		while (true) {
-			int c = read();
+			int c = (short) read();
 			switch (c) {
 			case ' ':
 			case '\n':
